backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pdf2image==1.17.0
Pillow==11.0.0
httpx==0.28.1
pydantic==2.10.3
